+++ internal/infrastructure/repositories/followsReposiroty/follow_repository.go
@@ -1,4 +1,4 @@
-package repository
+package followsReposiroty
 
 import (
 	"context"