+++ internal/infrastructure/repositories/tweetsRepository/tweet_repository.go
@@ -1,10 +1,10 @@
-package repository
+package tweetsRepository
 
 import (
 	"context"
 	"time"
-
 	"uala-challenge/internal/domain/tweets"
+	"uala-challenge/internal/infrastructure/repositories/followsReposiroty"
 
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
@@ -35,7 +35,7 @@ func (r *tweetRepository) CreateTweet(ctx context.Context, tweet *tweets.Tweet) 
 func (r *tweetRepository) GetTweetsByUserId(ctx context.Context, userId int) (tweetsList []tweets.Tweet, err error) {
 
 	// get the users that the user follows
-	followers, err := NewFollowRepository(r.collection.Database()).GetFollwersByUserId(ctx, userId)
+	followers, err := followsReposiroty.NewFollowRepository(r.collection.Database()).GetFollwersByUserId(ctx, userId)
 
 	if err != nil {
 		return