+++ CHANGELOG.md
@@ -9,9 +9,11 @@ _Documentation Update. Code Refactoring._
 
 ### Tweets
 * The domain component was refactored to use a individual package for tweets (update)
+* The repository component was refactored to use a individual package for tweets (update)
 
 ### Follows
 * The domain component was refactored to use a individual package for follows (update)
+* The repository component was refactored to use a individual package for follows (update)
 
 ### General Changes
 * Some tweaks and code revisions to use better practices (update)